requirements.txt
langchain
langgraph
langchain-google-genai
langchain-community
chromadb
python-dotenv
tiktoken
google-generativeai
streamlit